+++ app/react/components/Auth.tsx
@@ -1,5 +1,5 @@
-import React, { useEffect, useState } from "react";
-import app from "./base.tsx";
+import React, { useEffect, useState } from 'react';
+import app from '../../utils/base';
 
 export const AuthContext = React.createContext();
 
@@ -9,19 +9,19 @@ export const AuthProvider = ({ children }) => {
 
   useEffect(() => {
     app.auth().onAuthStateChanged((user) => {
-      setCurrentUser(user)
-      setPending(false)
+      setCurrentUser(user);
+      setPending(false);
     });
   }, []);
 
-  if(pending){
-    return <>Loading...</>
+  if (pending) {
+    return <>Loading...</>;
   }
 
   return (
     <AuthContext.Provider
       value={{
-        currentUser
+        currentUser,
       }}
     >
       {children}
