+++ app/react/components/PrivateRoute.tsx
@@ -1,0 +1,22 @@
+import React, { useContext } from 'react';
+import { Route, Redirect } from 'react-router-dom';
+import { AuthContext } from './Auth';
+import routes from '../constants/routes.json';
+
+const PrivateRoute = ({ component: RouteComponent, ...rest }) => {
+  const { currentUser } = useContext(AuthContext);
+  return (
+    <Route
+      {...rest}
+      render={(routeProps) =>
+        !!currentUser ? (
+          <RouteComponent {...routeProps} />
+        ) : (
+          <Redirect to={routes.LOGIN} />
+        )
+      }
+    />
+  );
+};
+
+export default PrivateRoute;
