--- app/react/pages/Login/Home.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import app from "./base";
-
-const Homee = () => {
-  return (
-    <>
-      <h1>Home</h1>
-      <button onClick={() => app.auth().signOut()}>Sign out</button>
-    </>
-  );
-};
-
-export default Homee;
