--- app/react/pages/Login/PrivateRoute.tsx
@@ -1,25 +1,0 @@
-import React, { useContext } from "react";
-import { Route, Redirect } from "react-router-dom";
-import { AuthContext } from "./Auth";
-import routes from '../../constants/routes.json';
-
-
-const PrivateRoute = ({ component: RouteComponent, ...rest }) => {
-  const {currentUser} = useContext(AuthContext);
-  console.log("I RENDERED SOMEWHERE")
-  return (
-    <Route
-      {...rest}
-      render={routeProps =>
-        !!currentUser ? (
-          <RouteComponent {...routeProps} />
-        ) : (
-          <Redirect to={routes.LOGIN} />
-        )
-      }
-    />
-  );
-};
-
-
-export default PrivateRoute
